refactor(plugin): cleanup, "de-mixin"-ification
