+++ src/cargo-install.js
@@ -4,7 +4,6 @@ const logger = require('@parcel/logger');
 const { proc } = require('./helpers');
 
 const installed = {};
-
 const isInstalled = cmdOrBin => installed[cmdOrBin];
 
 function cargoInstall(cmd, bin) {
