+++ src/helpers.js
@@ -1,3 +1,4 @@
+const path = require('path');
 const { execFile, spawn } = require('child_process');
 const { promisify } = require('util');
 
@@ -53,4 +54,21 @@ function proc(bin, args, opts) {
   });
 }
 
-module.exports = { exec, proc };
+function rel(from, to) {
+  let relativePath = path.relative(from, to);
+
+  if (relativePath[0] !== '.') {
+    relativePath = `./${relativePath}`;
+  }
+
+  return relativePath.replace('\\', '/');
+}
+
+function* matches(regex, str) {
+  let match;
+  while ((match = regex.exec(str)) !== null) {
+    yield match;
+  }
+}
+
+module.exports = { exec, proc, rel, matches };
